Stock retrieval from IBD50 (investors.com)
Retrieves stock list from investors.com for IBD50 for the week.  This is
a list of the best 50 stocks of the week.  Price pattern recognition
code as well.  There is also market specific code in here that can be
deleted.  For example, there is code that calculates distribution days
and follow throughs.  We already have that in our Market's data so that
can be ignored.